refactor: move ImageRequest and UnfairLock to separate files
